Improve postage algorithm
Before it was possible that no combination of stamps was found due to the perfect combination not being available and no larger stamp being available.
If that's the case it will now sum up the largest stamps in order to try to match the required postage.
This is basically a greedy implementation. If this should be further optimized you should look into the knapsack problem.